aws/microservices/user-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic[email]==2.5.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
prometheus-client==0.19.0
pydantic-settings==2.1.0
PyJWT==2.8.0 